Make getTestData() friendly with interrupt drive system tasks.
getTestData() now takes data from the sys data structure instead of reading directly from sensors.
Could not migrate code that tests the mux, as it requires direct reads and writes to and from the mux and there is no equivalent system to do so in the sys global structure. Commented code out and made it return 0.